+++ axum-core/src/response/headers.rs
@@ -7,7 +7,6 @@ use http::{
 };
 use http_body::{Empty, Full};
 use std::{convert::TryInto, fmt};
-use tower::util::Either;
 
 /// A response with headers.
 ///
@@ -146,6 +145,11 @@ where
     }
 }
 
+enum Either<A, B> {
+    A(A),
+    B(B),
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -172,7 +176,7 @@ mod tests {
         let res = (Headers(vec![("user-agent", "axum")]), "foo").into_response();
 
         assert_eq!(res.headers()["user-agent"], "axum");
-        let body = hyper::body::to_bytes(res.into_body())
+        let body = crate::body::to_bytes(res.into_body())
             .now_or_never()
             .unwrap()
             .unwrap();
@@ -190,7 +194,7 @@ mod tests {
 
         assert_eq!(res.headers()["user-agent"], "axum");
         assert_eq!(res.status(), StatusCode::NOT_FOUND);
-        let body = hyper::body::to_bytes(res.into_body())
+        let body = crate::body::to_bytes(res.into_body())
             .now_or_never()
             .unwrap()
             .unwrap();